ms-mcp-server/requirements.txt
fastapi-mcp==0.2.0
uvicorn==0.34.1
FastMCP==2.2.0
pandas==2.2.3
joblib==1.5.1
scikit-learn==1.6.1